% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actualiza_ubicaciones_genericas_ggmap.R
\name{actualiza_ubicaciones_genericas_ggmap}
\alias{actualiza_ubicaciones_genericas_ggmap}
\title{Actualiza y genera el listado de ubicaciones genericas}
\usage{
actualiza_ubicaciones_genericas_ggmap(loc_genericas,
  ubicaciones_genericas_path = "ubicaciones_genericas_ggmap.csv",
  apikey = readLines("apikey.txt"))
}
\arguments{
\item{loc_genericas}{Las ubicaciones genericas a ser buscadas}

\item{ubicaciones_genericas_path}{El path hacia el archivo de ubicaciones genericas. Si no existe, se creará en ese lugar el archivo nuevo.}

\item{apikey}{La clave api para las busquedas en google. Por default, espera que se encuentre en un archivo llamado "apikey.txt"}
}
\value{
Un data frame con las ubicaciones genericas cargadas
}
\description{
Esta función actualiza y carga la base de datos de ubicaciones genericas de google maps
}
