% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actualiza_ubicaciones_genericas_osm.R
\name{actualiza_ubicaciones_genericas_osm}
\alias{actualiza_ubicaciones_genericas_osm}
\title{Actualiza las ubicaciones genericas obtenidas por OSM}
\usage{
actualiza_ubicaciones_genericas_osm(loc_genericas,
  ubicaciones_genericas_path = "ubicaciones_genericas_osm.csv",
  timeout = 5, verbose = T)
}
\arguments{
\item{loc_genericas}{Las ubicaciones genericas a ser buscadas}

\item{ubicaciones_genericas_path}{El path hacia el archivo de ubicaciones genericas. Si no existe, se creará en ese lugar el archivo nuevo.}
}
\value{
Un data frame con las ubicaciones genericas cargadas
}
\description{
Esta función actualiza y carga la base de datos de ubicaciones genericas de OSM
}
