% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asocia_hospitales_data_a_hospitales_refes.R
\name{asocia_hospitales_data_a_hospitales_refes}
\alias{asocia_hospitales_data_a_hospitales_refes}
\title{Asocia clinicas con un archivo de referencia mediante asociación de texto}
\usage{
asocia_hospitales_data_a_hospitales_refes(data_ggmap, hospitales,
  hospi_name_column = "ESTAB_CLINICA")
}
\arguments{
\item{data_ggmap}{El dataset con las ubicaciones obtenidas por ggmap}

\item{hospitales}{Objecto sf conteniendo las ubicaciones de los hospitales.}

\item{hospi_name_column}{La columna que contiene el nombre de las clinicas en data_ggmap}
}
\value{
Una lista con las filas de hospitales, nombrada con los hospitales en data_ggmap
}
\description{
Esta función auxiliar encuentra las filas asociadas en el shp hospitales a los hospitales en data_ggmap
}
