% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistencia_columna.R
\name{calcula_consistencia_columna}
\alias{calcula_consistencia_columna}
\title{Verifica consistencia entre la ubicación y la provincia esperada}
\usage{
calcula_consistencia_columna(data_ggmap, data_osm, row_ggmap, row_osm,
  na.value, columna = "LAT_LON_EN_PROVINCIA")
}
\arguments{
\item{data_ggmap}{El dataset con las ubicaciones obtenidas por ggmap}

\item{data_osm}{El dataset con las ubicaciones obtenidas por OSM}

\item{row_ggmap}{La fila de data_ggmap a ser considerada}

\item{row_osm}{Las filas de data_osm asociadas a row_ggmap}

\item{na.value}{Como consideran en terminos de puntaje aquellos checkeos que hayan dado NA. Por default, es FALSE}
}
\value{
Un vector con TRUE y FALSE según si pasó o no el criterio. En caso de ser NA se asigna na.value
}
\description{
Esta funcion chequea la consistencia con el testeo presentado en una columna.
}
