% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistencia_distancia.R
\name{calcula_consistencia_distancia}
\alias{calcula_consistencia_distancia}
\title{Calcula el puntaje de consistencia de distancia entre las ubicaciones encontradas.}
\usage{
calcula_consistencia_distancia(data_ggmap, data_osm, row_ggmap, row_osm,
  lonlat_columns_ggmap, lonlat_columns_osm, id_column, max_dist)
}
\arguments{
\item{data_ggmap}{El dataset con las ubicaciones obtenidas por ggmap}

\item{data_osm}{El dataset con las ubicaciones obtenidas por OSM}

\item{row_ggmap}{La fila de data_ggmap a ser considerada}

\item{row_osm}{Las filas de data_osm asociadas a row_ggmap}

\item{lonlat_columns_ggmap}{Las columnas en las que se encuentras las ubicaciones lon y lat en data_ggmap}

\item{lonlat_columns_osm}{Las columnas en las que se encuentras las ubicaciones lon y lat en data_osm}

\item{id_column}{Columna para vincular data_ggmap con data_osm. Por default IDEVENTOCASO}

\item{max_dist}{La máxima distancia aceptable puede estar una ubicacion de las otras}
}
\value{
Un vector de longitud igual al numero de filas considerado, indicando lejos de cuantos puntos se encontraba cada ubicacion.
}
\description{
Esta funcion revisa que sean consistentes los puntos ubicados por ggmap y tmaptools
}
