% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_departamentos_ggmap.R
\name{check_departamentos_ggmap}
\alias{check_departamentos_ggmap}
\title{Comprueba consistencia entre los departamentos y las posiciones georreferenciadas (GOOGLE)}
\usage{
check_departamentos_ggmap(data_ggmap, departamentos,
  depto_id_column = "ID_DEPTO_INDEC_RESIDENCIA",
  lonlat_columns = c("LON_RESIDENCIA", "LAT_RESIDENCIA"),
  verbose = TRUE)
}
\arguments{
\item{data_ggmap}{El data frame con ubicaciones, y provincias asociadas}

\item{departamentos}{El objeto SF que contiene los poligonos de los departamentos. Se asume que tiene una columna "link" con los ids del indec del departamento.}

\item{verbose}{Boleano ¿Debe imprimirse el progreso? Por default TRUE}

\item{depto_id_colum}{La columna en la que está el id del departamento en data_ggmap}

\item{lonlat_column}{Las columnas lon y lat en las que están las ubicaciones del caso}
}
\value{
Un data frame identico a data_ggmap, pero que también contiene una columna con TRUE,FALSE,NA segun cumpla o no el criterio, o no se haya podido checkear.
}
\description{
Esta función toma un data frame, y un objeto sf con los departamentos, y checkea que las ubicaciones coincidan con los departamentos esperados
}
