% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_provincia_ggmap.R
\name{check_provincia_ggmap}
\alias{check_provincia_ggmap}
\title{Comprueba consistencia entre la provincia y las posiciones georreferenciadas (GOOGLE)}
\usage{
check_provincia_ggmap(data_ggmap, provincias,
  prov_id_column = "ID_PROV_INDEC_RESIDENCIA",
  lonlat_columns = c("LON_RESIDENCIA", "LAT_RESIDENCIA"), verbose = T,
  crs = 22185)
}
\arguments{
\item{data_ggmap}{El data frame con ubicaciones, y provincias asociadas}

\item{provincias}{El objeto SF que contiene los poligonos de las provincias. Se asume que tiene una columna "link" con los ids del indec de la provincia.}

\item{verbose}{Boleano ¿Debe imprimirse el progreso? Por default TRUE}

\item{crs}{Sistema coordenado de referencia sobre el cual buscar las intersecciones. Por default 22185}

\item{prov_id_colum}{La columna en la que está el id de la provincia en data_ggmap}

\item{lonlat_column}{Las columnas lon y lat en las que están las ubicaciones del caso}
}
\value{
Un data frame identico a data_ggmap, pero que también contiene una columna con TRUE,FALSE,NA segun cumpla o no el criterio, o no se haya podido checkear.
}
\description{
Esta función toma un data frame, y un objeto sf con las provincias, y checkea que las ubicaciones coincidan con las provincias esperadas
}
