% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ubicacion_generica_ggmap.R
\name{check_ubicacion_generica_ggmap}
\alias{check_ubicacion_generica_ggmap}
\title{Corrobora las ubicaciones genericas de GOOGLE}
\usage{
check_ubicacion_generica_ggmap(data_ggmap, ubicaciones_genericas_ggmap,
  campos_genericos1, campos_genericos2,
  lonlat_columns = c("LON_RESIDENCIA", "LAT_RESIDENCIA"),
  texto_a_eliminar = c("*SIN DATO* (*SIN DATO*)", "NULL"), verbose = T,
  invertir = F)
}
\arguments{
\item{data_ggmap}{El dataset con las ubicaciones obtenidas por ggmap}

\item{ubicaciones_genericas_ggmap}{El dataset con las ubicaciones genericas para ggmap}

\item{campos_genericos1}{Los campos a partir de los cuales reconstruir la ubicación generica}

\item{campos_genericos2}{Los campos a partir de los cuales reconstruir la ubicación generica, en una segunda versión}

\item{texto_a_eliminar}{El texto a ser eliminado en la construcción de las ubicaciones genericas. Por default c('*SIN DATO* (*SIN DATO*)','NULL')}

\item{verbose}{Boleano ¿Debe imprimirse el progreso? Por default TRUE}

\item{invertir}{Dado que la lógica de construccion de busqueda es Prefijo-Campo1-Campo2-...-CampoN, con esta opción se puede invertir la construccion a CampoN-...-Campo1-Prefijo. Por default, FALSE}

\item{lolat_columns}{Las columnas en las que se encuentras las ubicaciones lon y lat}
}
\value{
El mismo data frame, pero con una columna nueva que indica si pasó el check
}
\description{
La funcion genera en el archivo de datos una columna indicando si la ubicacion lonlat de cada fila es generica o no.
}
