% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ubicacion_generica_ggmap_y_osm.R
\name{check_ubicacion_generica_ggmap_y_osm}
\alias{check_ubicacion_generica_ggmap_y_osm}
\title{Comprueba que las ubicaciones obtenidas no sean genericas}
\usage{

  check_ubicacion_generica_ggmap_y_osm(inputArchivo = "bases/ejemplo/ejemplo.csv",
  ubicaciones_genericas_ggmap_path = "ubicaciones_genericas/ubicaciones_genericas_ggmap.csv",
  ubicaciones_genericas_osm_path = "ubicaciones_genericas/ubicaciones_genericas_osm.csv",
  campos_genericos1 = c("Localidad", "Calle", "Número"),
  campos_genericos2 = c("Partido", "Localidad", "Calle", "Número"),
  prefijo = "ARGENTINA", id_column = "IDEVENTOCASO",
  apikey = readLines("apikey.txt"), write.it = T, verbose = T,
  timeout = 5, texto_a_eliminar = c("*SIN DATO* (*SIN DATO*)", "NULL"),
  lonlat_columns_ggmap = c("LON_RESIDENCIA", "LAT_RESIDENCIA"),
  lonlat_columns_osm = c("LON_RESIDENCIA_OSM", "LAT_RESIDENCIA_OSM"))
}
\arguments{
\item{inputArchivo}{El path al archivo de datos}

\item{campos_genericos1}{Los campos del dataset en los cuales se encuentra la dirección. Deben ser indicados de mayor a menor, de país a numero. Por default c("Localidad","Calle","Número")}

\item{campos_genericos2}{Los campos del dataset en los cuales se encuentra la dirección. Deben ser indicados de mayor a menor, de país a numero. Por default c("Localidad","Calle","Número")}

\item{prefijo}{Si se desea, se puede agregar un prefijo al texto de busqueda}

\item{apikey}{La clave api de google. Por default, busca apikey.txt}

\item{write.it}{Debe escribirse el resultado? Por default TRUE}

\item{verbose}{Boleano, ¿debe indicarse el resultado? Por default TRUE}

\item{timeout}{La cantidad de tiempo máxima a esperar por busqueda en OSM}
}
\value{
Devuelve una lista con ambos datasets
}
\description{
Esta función calcula la georreferenciación de los datos mediante ggmap y osm

Esta función calcula la georreferenciación de los datos mediante osm y devuelve el data frame, con columnas LAT_RESIDENCIA_OSM y LON_RESIDENCIA_OSM.
}
