% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ubicacion_generica_osm.R
\name{check_ubicacion_generica_osm}
\alias{check_ubicacion_generica_osm}
\title{Compruebas las ubicaciones genericas de OSM}
\usage{
check_ubicacion_generica_osm(data_osm, data_ggmap,
  ubicaciones_genericas_osm, campos_genericos1, campos_genericos2,
  lonlat_columns = c("LON_RESIDENCIA_OSM", "LAT_RESIDENCIA_OSM"),
  texto_a_eliminar = c("*SIN DATO* (*SIN DATO*)", "NULL"), verbose = T)
}
\arguments{
\item{data_osm}{El dataset con las ubicaciones obtenidas por osm}

\item{data_ggmap}{El dataset con las ubicaciones obtenidas por ggmap}

\item{ubicaciones_genericas_osm}{El dataset con las ubicaciones genericas para osm}

\item{campos_genericos1}{Los campos a partir de los cuales reconstruir la ubicación generica}

\item{campos_genericos2}{Los campos a partir de los cuales reconstruir la ubicación generica, en una segunda versión}

\item{verbose}{Boleano ¿Debe imprimirse el progreso? Por default TRUE}

\item{lolat_columns}{Las columnas en las que se encuentras las ubicaciones lon y lat}

\item{text_a_eliminar}{El texto a ser eliminado en la construcción de las ubicaciones genericas. Por default c('*SIN DATO* (*SIN DATO*)','NULL')}
}
\value{
El mismo data frame, pero con una columna nueva que indica si pasó el check
}
\description{
La funcion genera en el archivo de datos una columna indicando si la ubicacion lonlat de cada fila es generica o no.
}
