% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkea_clinica_ggmap.R
\name{checkea_clinica_ggmap}
\alias{checkea_clinica_ggmap}
\title{Comprueba cercanía con el punto de atencion (GOOGLE)}
\usage{
checkea_clinica_ggmap(data_ggmap, hospis_datos_shp, hospitales,
  max_distancia = 200 * 1000, lonlat_columns = c("LON_RESIDENCIA",
  "LAT_RESIDENCIA"), hospi_name_column = "ESTAB_CLINICA", verbose = T)
}
\arguments{
\item{data_ggmap}{El dataset con las ubicaciones obtenidas por ggmap}

\item{hospis_datos_shp}{Un vector con nombres en los establecimientos clinicos de data_ggmap, y valores en las filas correspondientes en hospitales}

\item{hospitales}{Un objeto SF con los puntos de los hospitales}

\item{max_distancia}{La máxima distancia aceptable a la que puede estar el hospital, en metros}

\item{hospi_name_column}{La columna que contiene el nombre de las clinicas en data_ggmap}

\item{verbose}{Boleano ¿Debe imprimirse el progreso? Por default TRUE}

\item{lolat_columns}{Las columnas en las que se encuentras las ubicaciones lon y lat}
}
\value{
El mismo data frame, pero con una columna nueva que indica si pasó el check, y además la posición lonlat de la clinica
}
\description{
La funcion genera en el archivo de datos una columna indicando si la ubicacion lonlat de cada fila esta lo suficientemente cerca del hospital o no.
}
