% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkea_clinica_ggmap_y_osm.R
\name{checkea_clinica_ggmap_y_osm}
\alias{checkea_clinica_ggmap_y_osm}
\title{Comprueba cercanía con el punto de atencion}
\usage{
checkea_clinica_ggmap_y_osm(inputArchivo = "bases/ejemplo/ejemplo.csv",
  hospitales_path = "shp/hospitales/hospitales.shp",
  hospi_name_column = "ESTAB_CLINICA",
  lonlat_columns_ggmap = c("LON_RESIDENCIA", "LAT_RESIDENCIA"),
  lonlat_columns_osm = c("LON_RESIDENCIA_OSM", "LAT_RESIDENCIA_OSM"),
  max_distancia = 200 * 1000, id_column = "IDEVENTOCASO",
  write.it = T, verbose = T)
}
\arguments{
\item{inputArchivo}{Path a partir del cual construir los paths para ggmap y osm. El archivo que fue usado para construirlos.}

\item{hospitales_path}{Path al archivo shp de las provincias. Por default 'shp/provincias/provincias.shp'}

\item{hospi_name_column}{Columna con el id de las provincias, por default 'ID_PROVINCIA'}

\item{lonlat_columns_ggmap}{Columnas con las ubicaciones lon y lat de las filas en data_ggmap. Por default c('LON_RESIDENCIA','LAT_RESIDENCIA')}

\item{lonlat_columns_osm}{Columnas con las ubicaciones lon y lat de las filas en data_osm. Por default c('LON_RESIDENCIA_OSM','LAT_RESIDENCIA_OSM')}

\item{id_column}{Columna con los ids en el archivo OSM. Por default, 'IDEVENTOCASO'}

\item{write.it}{Se deben escribir los archivos? Por default TRUE}

\item{verbose}{Booleano ¿debe imprimir el progreso? Default TRUE}
}
\value{
Una lista con ambos data frames.
}
\description{
Realiza la validación de las posiciones en base a la provincia indicada.
}
