% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkea_departamentos_ggmap_y_osm.R
\name{checkea_departamentos_ggmap_y_osm}
\alias{checkea_departamentos_ggmap_y_osm}
\title{Comprueba consistencia entre los departamentos y las posiciones georreferenciadas}
\usage{

  checkea_departamentos_ggmap_y_osm(inputArchivo = "bases/ejemplo/ejemplo.csv",
  departamentos_path = "shp/departamentos/censo_por_departamento_polígonos.shp",
  depto_id_column = "ID_DEPTO",
  lonlat_columns_ggmap = c("LON_RESIDENCIA", "LAT_RESIDENCIA"),
  lonlat_columns_osm = c("LON_RESIDENCIA_OSM", "LAT_RESIDENCIA_OSM"),
  crs = 22185, id_column = "IDEVENTOCASO", verbose = T,
  write.it = T)
}
\arguments{
\item{inputArchivo}{Path a partir del cual construir los paths para ggmap y osm. El archivo que fue usado para construirlos.}

\item{departamentos_path}{Path al archivo shp de las provincias. Por default 'shp/provincias/provincias.shp'}

\item{depto_id_column}{Columna con el id de las provincias, por default 'ID_PROVINCIA'}

\item{lonlat_columns_ggmap}{Columnas con las ubicaciones lon y lat de las filas en data_ggmap. Por default c('LON_RESIDENCIA','LAT_RESIDENCIA')}

\item{lonlat_columns_osm}{Columnas con las ubicaciones lon y lat de las filas en data_osm. Por default c('LON_RESIDENCIA_OSM','LAT_RESIDENCIA_OSM')}

\item{crs}{crs usado para buscar la interseccion. Por default 22185}

\item{id_column}{Columna con los ids en el archivo OSM. Por default, 'IDEVENTOCASO'}

\item{verbose}{Booleano ¿debe imprimir el progreso? Default TRUE}

\item{write.it}{Se deben escribir los archivos? Por default TRUE}
}
\value{
Una lista con ambos data frames.
}
\description{
Realiza la validación de las posiciones en base a la provincia indicada.
}
