% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combina_bases.R
\name{combina_bases}
\alias{combina_bases}
\title{Combina dos bases}
\usage{
combina_bases(inputArchivo1, inputArchivo2,
  outputArchivo = paste(sub(".csv", "", inputArchivo1), inputArchivo2,
  sep = "_pegado_"), id_column = "IDEVENTOCASO", ...)
}
\arguments{
\item{inputArchivo1}{El primer archivo, como path hasta el archivo.}

\item{inputArchivo2}{El segundo archivo, como path hasta el archivo}

\item{outputArchivo}{El archivo de salida en el cual escribir el resultado}

\item{id_column}{La columna en base a la cual matchear casos en ambos datasets}

\item{...}{Se pasa a read.csv}
}
\value{
El dataset combinado, como data.frame.
}
\description{
Esta función combina dos datasets en uno, poniendo primicia en el primero. Casos que estén repetidos en el segundo seran removidos.
}
