% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compara_GGMAP_OSM.R
\name{compara_GGMAP_OSM}
\alias{compara_GGMAP_OSM}
\title{Compara los resultados obtenidos con GGMAP y OSM y elige el mejor resultado}
\usage{
compara_GGMAP_OSM(inputArchivo = "bases/Base_p_ariel_y_yamila.csv",
  peso_hospital = 10, peso_generico = 1, peso_provincia = 1,
  peso_departamento = 1, peso_distancia = 1, puntaje_de_corte = -5,
  lonlat_columns_ggmap = c("LON_RESIDENCIA", "LAT_RESIDENCIA"),
  lonlat_columns_osm = c("LON_RESIDENCIA_OSM", "LAT_RESIDENCIA_OSM"),
  max_dist = 1000, id_column = "IDEVENTOCASO", na.value = FALSE,
  write.it = T, verbose = T, mapa_check = F)
}
\arguments{
\item{inputArchivo}{Path a partir del cual construir los paths para ggmap y osm. El archivo que fue usado para construirlos.}

\item{peso_hospital}{El peso asociado al checkeo.}

\item{peso_provincia}{El peso asociado a los distintos checkeos.}

\item{peso_distancia}{El peso asociado a los distintos checkeos.}

\item{puntaje_de_corte}{El puntaje a partir del cual se indica la ubicacion de la clinica en vez de la georreferenciada. Además, las ubicaciones BAJAS son las que quedan por debajo de este valor}

\item{lonlat_columns_ggmap}{Las columnas en las que se encuentras las ubicaciones lon y lat en data_ggmap}

\item{lonlat_columns_osm}{Las columnas en las que se encuentras las ubicaciones lon y lat en data_osm}

\item{max_dist}{La máxima distancia aceptable puede estar una ubicacion de las otras}

\item{id_column}{Columna para vincular data_ggmap con data_osm. Por default IDEVENTOCASO}

\item{na.value}{Como consideran en terminos de puntaje aquellos checkeos que hayan dado NA. Por default, es FALSE}

\item{write.it}{Se deben escribir los archivos? Por default TRUE}

\item{verbose}{Boleano ¿Debe imprimirse el progreso? Por default TRUE}

\item{mapa_check}{Si es TRUE, guarda también un mapa generado con leaflet con los casos y colores según su calidad.}

\item{peso_hospital}{El peso asociado a los distintos checkeos.}
}
\value{
Un data frame con la ubicacion elegida para cada caso, indicando el puntaje que obtuvo y de donde proviene, así como el agregado de la ubicación de la clinica.
}
\description{
La funcion genera im data frame con las posiciones exitosas de comparar GGMAP y OSM. Al final, la ubicación propuesta es acompañada de una etiqueta según su confiabilidad sea BUENA, MEDIA o BAJA.
}
