% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encuentra_ubicacion_generica.R
\name{encuentra_ubicacion_generica_ggmap}
\alias{encuentra_ubicacion_generica_ggmap}
\title{Encuentra las ubicaciones latlon asociadas a un conjunto de datos}
\usage{
encuentra_ubicacion_generica_ggmap(data_ggmap, campos_genericos,
  ubicaciones_genericas_path, prefijo = "ARGENTINA",
  apikey = readLines("apikey.txt"))
}
\arguments{
\item{campos_genericos}{Los campos a ser pegados}

\item{ubicaciones_genericas_path}{El path al archivo con las ubicaciones genericas.}

\item{prefijo}{Un campo de texto a ser agregado previo al resto. Por default "ARGENTINA"}

\item{apikey}{La clave api de google. Por default, busca apikey.txt}

\item{datos}{Un data frame con los campos_a_domicilio presentes}

\item{texto_a_eliminar}{Texto a ser omitido en caso de ser encontrado (por comparación exacta). Por default c('*SIN DATO* (*SIN DATO*)','NULL',paste('COMUNA',1:20))}

\item{sep}{El separador a ser usado cuando se peguen los campos. Por default ", "}

\item{campoNumero}{Campo con el número de la calle}
}
\value{
El mismo dataset, pero con columnas LAT_GENERICA y LON_GENERICA
}
\description{
Esta función toma el data frame datos y ciertos campos del mismo, y encuentra sus ubicaciones genericas.
}
