% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genera_loc_domicilios.R
\name{genera_loc_domicilios}
\alias{genera_loc_domicilios}
\title{Genera la lista de domicilios para buscar}
\usage{
genera_loc_domicilios(datos, campos,
  texto_a_eliminar = c("*SIN DATO* (*SIN DATO*)", "NULL", paste("COMUNA",
  1:20)), prefijo = "ARGENTINA", sep = ", ", campoNumero = NULL,
  invertir = F)
}
\arguments{
\item{datos}{Un data frame con los campos_a_domicilio presentes}

\item{campos}{Los campos a ser pegados}

\item{prefijo}{Un campo de texto a ser agregado previo al resto. Por default "ARGENTINA"}

\item{sep}{El separador a ser usado cuando se peguen los campos. Por default ", "}

\item{campoNumero}{Campo con el número de la calle}

\item{invertir}{Dado que la lógica de construccion es Prefijo-Campo1-Campo2-...-CampoN, con esta opción se puede invertir la construccion a CampoN-...-Campo1-Prefijo. Por default, FALSE}

\item{texto_a_omitir}{Texto a ser omitido en caso de ser encontrado (por comparación exacta). Por default c('*SIN DATO* (*SIN DATO*)','NULL',paste('COMUNA',1:20))}
}
\value{
Un vector de campos para ser georreferenciados.
}
\description{
Esta función toma el data frame datos y ciertos campos del mismo, y los usa para generar un vector de strings para ser georreferenciados
}
