% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georreferencia_ggmap.R
\name{georreferencia_ggmap}
\alias{georreferencia_ggmap}
\title{Georreferencia usando ggmap}
\usage{
georreferencia_ggmap(inputArchivo = "bases/ejemplo/ejemplo.csv",
  campos_a_domicilio = c("Localidad", "Calle", "Número"),
  prefijo_domicilio = "ARGENTINA", apikey = readLines("apikey.txt"),
  sep = ",", write.it = T, verbose = T, invertir = F)
}
\arguments{
\item{inputArchivo}{El path al archivo de datos}

\item{campos_a_domicilio}{Los campos del dataset en los cuales se encuentra la dirección. Deben ser indicados de mayor a menor, de país a numero. Por default c("Localidad","Calle","Número")}

\item{prefijo_domicilio}{Si se desea, se puede agregar un prefijo al texto de busqueda}

\item{apikey}{La clave api de google. Por default, busca apikey.txt}

\item{sep}{El separador de columnas en inputArchivo}

\item{write.it}{Debe escribirse el resultado? Por default TRUE}

\item{verbose}{Boleano, ¿debe indicarse el resultado? Por default TRUE}

\item{invertir}{Dado que la lógica de construccion de busqueda es Prefijo-Campo1-Campo2-...-CampoN, con esta opción se puede invertir la construccion a CampoN-...-Campo1-Prefijo. Por default, FALSE}
}
\value{
Devuelve el dataset con las columnas agregadas.
}
\description{
Esta función calcula la georreferenciación de los datos mediante ggmap y devuelve el data frame, con columnas LAT_RESIDENCIA y LON_RESIDENCIA.
}
