% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georreferencia_osm.R
\name{georreferencia_osm}
\alias{georreferencia_osm}
\title{Georreferencia usando tmaptools}
\usage{
georreferencia_osm(inputArchivo = sub(".csv", "_recortados.csv",
  "bases/francoA/ListadoA.csv"), id_column = "IDEVENTOCASO",
  campos_a_domicilio = c("Localidad", "Calle", "Número"),
  campos_a_domicilio2 = c("Partido", "Localidad", "Calle", "Número"),
  prefijo_domicilio = "ARGENTINA", sep = ",", write.it = T,
  verbose = T, timeout = 5)
}
\arguments{
\item{inputArchivo}{El path al archivo de datos}

\item{campos_a_domicilio}{Los campos del dataset en los cuales se encuentra la dirección. Deben ser indicados de mayor a menor, de país a numero. Por default c("Localidad","Calle","Número")}

\item{campos_a_domicilio2}{Los campos del dataset en los cuales se encuentra la dirección. Deben ser indicados de mayor a menor, de país a numero. Por default c("Localidad","Calle","Número")}

\item{prefijo_domicilio}{Si se desea, se puede agregar un prefijo al texto de busqueda}

\item{sep}{El separador de columnas en inputArchivo}

\item{write.it}{Debe escribirse el resultado? Por default TRUE}

\item{verbose}{Boleano, ¿debe indicarse el resultado? Por default TRUE}
}
\value{
Devuelve el dataset con las columnas agregadas.
}
\description{
Esta función calcula la georreferenciación de los datos mediante osm y devuelve el data frame, con columnas LAT_RESIDENCIA_OSM y LON_RESIDENCIA_OSM.
}
