% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocesa_texto.R
\name{preprocesa_texto}
\alias{preprocesa_texto}
\title{Preprocesa el texto para hacer más exitosa la búsqueda.}
\usage{
preprocesa_texto(texto, remueve.post.entre = F)
}
\arguments{
\item{texto}{El vector con los campos a procesar}

\item{remueve.post.entre}{Si es igual a TRUE, remueve todo el texto que aparezca después de una expresión asociada a "ENTRE"}
}
\value{
El vector, ya preprocesado.
}
\description{
Esta función hace ciertos preprocesamientos básicos sobre el texto para que sea más fácilmente aceptado por los buscadores. Elimina caracteres especiales, remueve tildes, separa texto de números, y desabrevia la palabra número.
}
