% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recorta_bases.R
\name{recorta_bases}
\alias{recorta_bases}
\title{Recorta una base segun los casos presentes en otra}
\usage{
recorta_bases(inputArchivo1, inputArchivo2,
  outputArchivo = paste(sub(".csv", "", inputArchivo1), inputArchivo2,
  sep = "_recortado_"), columnas = "IDEVENTOCASO", ...)
}
\arguments{
\item{inputArchivo1}{El primer archivo, como path hasta el archivo.}

\item{inputArchivo2}{El segundo archivo, como path hasta el archivo}

\item{outputArchivo}{El archivo de salida en el cual escribir el resultado}

\item{columnas}{Las columnas mediante las cuales matchear los datasets}

\item{...}{Se pasa a read.csv}
}
\value{
El dataset recortado.
}
\description{
Esta función remueve casos repetidos del dataset 2, que esten presentes en el 1. Para esto, compara las columnas en columnas
}
