% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecciona_casos.R
\name{selecciona_casos}
\alias{selecciona_casos}
\title{Subselecciona una base segun los casos de una columna}
\usage{
selecciona_casos(dataset, columna, casos)
}
\arguments{
\item{dataset}{La base de datos}

\item{columna}{La columna en la que buscar casos}

\item{casos}{Los casos a retener de esa columna}
}
\value{
El dataset recortado.
}
\description{
Esta funcion recorta un dataset en base a una columna, quedandose con ciertos casos.
}
